♻️refactor: socket type 분리
