+++ src/types/api.ts
@@ -49,13 +49,3 @@ export interface BalanceGameSelectionsResponse {
   b: string[];
   c: string[];
 }
-
-export interface BalanceGameRoundResponse {
-  totalRounds: number;
-  currentRound: number;
-  startTime: string;
-  endTime: string;
-  q: string;
-  a: string;
-  b: string;
-}
