+++ src/types/websocket.ts
@@ -1,0 +1,11 @@
+/* Response */
+
+export interface BalanceGameRoundResponse {
+  totalRounds: number;
+  currentRound: number;
+  startTime: string;
+  endTime: string;
+  q: string;
+  a: string;
+  b: string;
+}
